+++ src/reference/toHtml.js
@@ -6,10 +6,10 @@ const toHtml = function(c, options) {
     if (options.links === true) {
       str = `<a href="${c.data.url}">${str}</a>`;
     }
-    return `<div class="citation">⌃ ${str} </div>`;
+    return `<div class="reference">⌃ ${str} </div>`;
   }
   if (c.data.encyclopedia) {
-    return `<div class="citation">⌃ ${c.data.encyclopedia}</div>`;
+    return `<div class="reference">⌃ ${c.data.encyclopedia}</div>`;
   }
   if (c.data.title) { //cite book, etc
     let str = c.data.title;
@@ -19,10 +19,10 @@ const toHtml = function(c, options) {
     if (c.data.first && c.data.last) {
       str += c.data.first + ' ' + c.data.last;
     }
-    return `<div class="citation">⌃ ${str}</div>`;
+    return `<div class="reference">⌃ ${str}</div>`;
   }
   if (c.inline) {
-    return `<div class="citation">⌃ ${c.inline.html()}</div>`;
+    return `<div class="reference">⌃ ${c.inline.html()}</div>`;
   }
   return '';
 };