+++ changelog.md
@@ -117,3 +117,4 @@
 * hide object data in console.logs
 * move ALL image urls from `upload.wikimedia.org/wikipedia/commons` to `wikipedia.org/wiki/Special:Redirect/file/` via #86
 * image captions are now Sentence objects
+* rename citation → reference internally, and in json output